# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Coopuertos",
        "description": "API REST para el sistema Coopuertos - Gestión de conductores, vehículos y propietarios",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor API v1"
        }
    ],
    "paths": {
...
        "/api/v1/conductores": {
            "get": {
                "tags": [
                    "Conductores"
                ],
                "summary": "Listar conductores",
                "description": "Retorna una lista paginada de conductores con filtros opcionales",
                "operationId": "ab2d8caf0998204767665b30d845b4ce",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de conductores"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Conductores"
                ],
                "summary": "Crear conductor",
                "description": "Crea un nuevo conductor",
                "operationId": "b6b362bccfe151f99d6aed388bd75b07",
                "responses": {
                    "201": {
                        "description": "Conductor creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/conductores/{id}": {
            "get": {
                "tags": [
                    "Conductores"
                ],
                "summary": "Mostrar conductor",
                "description": "Retorna la información de un conductor específico",
                "operationId": "53df2e0ee0393e9cd732e4fc9e8b5870",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del conductor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del conductor"
                    },
                    "404": {
                        "description": "Conductor no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Conductores"
                ],
                "summary": "Actualizar conductor",
                "description": "Actualiza la información de un conductor",
                "operationId": "0ddbf2839fabe3406027a2ff2a09cb80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del conductor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conductor actualizado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Conductor no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Conductores"
                ],
                "summary": "Eliminar conductor",
                "description": "Elimina un conductor del sistema",
                "operationId": "922c1d4bebec1238ecb419f7edc43293",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del conductor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conductor eliminado exitosamente"
                    },
                    "404": {
                        "description": "Conductor no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/conductores/search": {
            "get": {
                "tags": [
                    "Conductores"
                ],
                "summary": "Buscar conductores",
                "description": "Busca conductores por cédula, nombres, apellidos o número interno",
                "operationId": "9967584712f95aa621f90160bfb75fc7",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de búsqueda"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/conductores/{uuid}/public": {
            "get": {
                "tags": [
                    "Conductores"
                ],
                "summary": "Mostrar conductor público",
                "description": "Retorna la información pública de un conductor por UUID (sin autenticación)",
                "operationId": "50cd17c31a1841633cde8f01137b12bd",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID del conductor",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del conductor"
                    },
                    "404": {
                        "description": "Conductor no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación de Laravel Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}